consolidate Asset model in common
